+++ vscode/webviews/chat/ChatMessageContent/create-buttons.ts
@@ -1,0 +1,297 @@
+import {
+    CheckCodeBlockIcon,
+    CopyCodeBlockIcon,
+    EllipsisIcon,
+    InsertCodeBlockIcon,
+    SaveCodeBlockIcon,
+} from '../../icons/CodeBlockActionIcons'
+
+import type { CodyTaskState } from '../../../src/non-stop/state'
+import styles from './ChatMessageContent.module.css'
+import type { PriorHumanMessageInfo } from '../cells/messageCell/assistant/AssistantMessageCell'
+import type { CodeBlockActionsProps } from './ChatMessageContent'
+
+export function createButtons(
+    preText: string,
+    copyButtonOnSubmit?: CodeBlockActionsProps['copyButtonOnSubmit'],
+    insertButtonOnSubmit?: CodeBlockActionsProps['insertButtonOnSubmit']
+): HTMLElement {
+    const container = document.createElement('div')
+    container.className = styles.buttonsContainer
+
+    if (!copyButtonOnSubmit) {
+        return container
+    }
+
+    // The container will contain the buttons and the <pre> element with the code.
+    // This allows us to position the buttons independent of the code.
+    const buttons = document.createElement('div')
+    buttons.className = styles.buttons
+
+    const codeBlockActions = {
+        copy: copyButtonOnSubmit,
+        insert: insertButtonOnSubmit,
+    }
+
+    const copyButton = createCodeBlockActionButton(
+        'copy',
+        preText,
+        'Copy Code',
+        CopyCodeBlockIcon,
+        codeBlockActions
+    )
+    buttons.append(copyButton)
+
+    // The insert buttons only exists for IDE integrations
+    if (insertButtonOnSubmit) {
+        buttons.append(
+            createCodeBlockActionButton(
+                'insert',
+                preText,
+                'Insert Code at Cursor',
+                InsertCodeBlockIcon,
+                codeBlockActions
+            )
+        )
+
+        buttons.append(
+            createCodeBlockActionButton(
+                'new',
+                preText,
+                'Save Code to New File...',
+                SaveCodeBlockIcon,
+                codeBlockActions
+            )
+        )
+    }
+
+    container.append(buttons)
+
+    return container
+}
+
+export function createButtonsExperimentalUI(
+    preText: string,
+    humanMessage: PriorHumanMessageInfo | null,
+    fileName?: string,
+    copyButtonOnSubmit?: CodeBlockActionsProps['copyButtonOnSubmit'],
+    smartApply?: CodeBlockActionsProps['smartApply'],
+    smartApplyId?: string,
+    smartApplyState?: CodyTaskState
+): HTMLElement {
+    const container = document.createElement('div')
+    container.className = styles.buttonsContainer
+    if (!copyButtonOnSubmit) {
+        return container
+    }
+
+    const buttons = document.createElement('div')
+    buttons.className = styles.buttons
+
+    if (smartApplyState === 'Applied' && smartApplyId) {
+        const acceptButton = createAcceptButton(smartApplyId, smartApply)
+        const discardButton = createDiscardButton(smartApplyId, smartApply)
+        buttons.append(acceptButton, discardButton)
+    } else {
+        const copyButton = createCopyButton(preText, copyButtonOnSubmit)
+        buttons.append(copyButton)
+
+        if (smartApply) {
+            const applyButton = createApplyButton(
+                preText,
+                humanMessage,
+                smartApply,
+                smartApplyState,
+                fileName
+            )
+            buttons.append(applyButton)
+        }
+
+        const actionsDropdown = createActionsDropdown(preText)
+        buttons.append(actionsDropdown)
+    }
+
+    container.append(buttons)
+
+    if (fileName) {
+        const fileNameContainer = document.createElement('div')
+        fileNameContainer.className = styles.fileNameContainer
+        fileNameContainer.textContent = getFileName(fileName)
+        container.append(fileNameContainer)
+    }
+
+    return container
+}
+
+/**
+ * Creates a button to perform an action on a code block.
+ * @returns The button element.
+ */
+function createCodeBlockActionButton(
+    type: 'copy' | 'insert' | 'new',
+    text: string,
+    title: string,
+    iconSvg: string,
+    codeBlockActions: {
+        copy: CodeBlockActionsProps['copyButtonOnSubmit']
+        insert?: CodeBlockActionsProps['insertButtonOnSubmit']
+    }
+): HTMLElement {
+    const button = document.createElement('button')
+
+    const className = type === 'copy' ? styles.copyButton : styles.insertButton
+
+    button.innerHTML = iconSvg
+    button.title = title
+    button.className = className
+
+    if (type === 'copy') {
+        button.addEventListener('click', () => {
+            button.innerHTML = CheckCodeBlockIcon
+            navigator.clipboard.writeText(text).catch(error => console.error(error))
+            button.className = className
+            codeBlockActions.copy(text, 'Button')
+            setTimeout(() => {
+                button.innerHTML = iconSvg
+            }, 5000)
+
+            // Log for `chat assistant response code buttons` e2e test.
+            console.log('Code: Copy to Clipboard', text)
+        })
+    }
+
+    const insertOnSubmit = codeBlockActions.insert
+    if (!insertOnSubmit) {
+        return button
+    }
+
+    switch (type) {
+        case 'insert':
+            button.addEventListener('click', () => insertOnSubmit(text, false))
+            break
+        case 'new':
+            button.addEventListener('click', () => insertOnSubmit(text, true))
+            break
+    }
+
+    return button
+}
+
+function createCopyButton(
+    preText: string,
+    onCopy: CodeBlockActionsProps['copyButtonOnSubmit']
+): HTMLElement {
+    const button = document.createElement('button')
+    button.innerHTML = 'Copy'
+    button.className = styles.button
+
+    const iconContainer = document.createElement('div')
+    iconContainer.className = styles.iconContainer
+    iconContainer.innerHTML = CopyCodeBlockIcon
+    button.prepend(iconContainer)
+
+    button.addEventListener('click', () => {
+        iconContainer.innerHTML = CheckCodeBlockIcon
+        iconContainer.className = styles.iconContainer
+        button.innerHTML = 'Copied'
+        button.className = styles.button
+        button.prepend(iconContainer)
+
+        navigator.clipboard.writeText(preText).catch(error => console.error(error))
+        onCopy(preText, 'Button')
+        setTimeout(() => {
+            // Reset the icon to the original.
+            iconContainer.innerHTML = CopyCodeBlockIcon
+            iconContainer.className = styles.iconContainer
+            button.innerHTML = 'Copy'
+            button.className = styles.button
+            button.prepend(iconContainer)
+        }, 5000)
+
+        // Log for `chat assistant response code buttons` e2e test.
+        console.log('Code: Copy to Clipboard', preText)
+    })
+
+    return button
+}
+
+function createApplyButton(
+    preText: string,
+    humanMessage: PriorHumanMessageInfo | null,
+    smartApply: CodeBlockActionsProps['smartApply'],
+    applyState?: CodyTaskState,
+    fileName?: string
+): HTMLElement {
+    const button = document.createElement('button')
+    button.className = styles.button
+
+    switch (applyState) {
+        case 'Working':
+            button.innerHTML = 'Applying...'
+            button.disabled = true
+            break
+        case 'Error':
+            button.innerHTML = 'Error'
+            button.disabled = true
+            break
+        default:
+            button.innerHTML = 'Apply'
+            button.addEventListener('click', () => {
+                const smartApplyId = preText.trim()
+                smartApply?.onSubmit(smartApplyId, preText, humanMessage?.text, fileName)
+            })
+    }
+
+    return button
+}
+
+function createAcceptButton(id: string, smartApply: CodeBlockActionsProps['smartApply']): HTMLElement {
+    const button = document.createElement('button')
+    button.className = styles.button
+    button.innerHTML = 'Accept'
+    button.addEventListener('click', () => {
+        smartApply?.onAccept(id)
+    })
+    return button
+}
+
+function createDiscardButton(id: string, smartApply: CodeBlockActionsProps['smartApply']): HTMLElement {
+    const button = document.createElement('button')
+    button.className = styles.button
+    button.innerHTML = 'Discard'
+    button.addEventListener('click', () => {
+        smartApply?.onReject(id)
+    })
+    return button
+}
+
+function createActionsDropdown(preText: string): HTMLElement {
+    const button = document.createElement('button')
+    button.innerHTML = EllipsisIcon
+    button.title = 'More Actions...'
+    button.className = styles.button
+
+    const vscodeContext = {
+        webviewSection: 'codeblock-actions',
+        preventDefaultContextMenuItems: true,
+        text: preText,
+    }
+
+    // Attach `data-vscode-context`, this is also provided when the commands are executed,
+    // so serves as a way for us to pass `vscodeContext.text` to each relevant command
+    button.setAttribute('data-vscode-context', JSON.stringify(vscodeContext))
+
+    button.addEventListener('click', event => {
+        event.preventDefault()
+        event.target?.dispatchEvent(
+            new MouseEvent('contextmenu', {
+                bubbles: true,
+                clientX: event.clientX,
+                clientY: event.clientY,
+            })
+        )
+        event.stopPropagation()
+    })
+
+    return button
+}
