+++ vscode/webviews/chat/ChatMessageContent/utils.ts
@@ -1,0 +1,3 @@
+function getFileName(filePath: string): string {
+    return filePath.split('/').pop() || filePath
+}
