+++ vscode/webviews/chat/cells/messageCell/assistant/AssistantMessageCell.tsx
@@ -16,7 +16,7 @@ import { type FunctionComponent, type RefObject, memo, useMemo } from 'react'
 import type { ApiPostMessage, UserAccountInfo } from '../../../../Chat'
 import { chatModelIconComponent } from '../../../../components/ChatModelIcon'
 import { Tooltip, TooltipContent, TooltipTrigger } from '../../../../components/shadcn/ui/tooltip'
-import { ChatMessageContent, type CodeBlockActionsProps } from '../../../ChatMessageContent'
+import { ChatMessageContent, type CodeBlockActionsProps } from '../../../ChatMessageContent/ChatMessageContent'
 import { ErrorItem, RequestErrorItem } from '../../../ErrorItem'
 import { type Interaction, editHumanMessage } from '../../../Transcript'
 import { FeedbackButtons } from '../../../components/FeedbackButtons'
